Added manager ability to delete team accounts in /equipe
Added an column named "Action" to the team table in /equipe, the column is visible only to a Manager account, selecting the "Supprimer" button in a row sends a post request to /equipe/supprimer-membre with the email address of the team member to be deleted, this email is sent named "deleteEmail" in the post request. the team member account deletion happens in ManagerController@deleteMember.