Added /client, and some other changes
Added listing of clients in a card format in /client (visible to managers and teleoperateurs).
Added the ability for managers to add clients.
Added client image uploading (ManagerController@imageUpload), images get saved to /public/images with a random name, the name of the file of each client's image is saved in table clients@image.
Routing preparations for client deletion using the route "/clients/supprimer-client" .
Changed placement of the "Ajouter un produit" button in /produits from the body to the header, aligning it with the design of other pages.