Added /dashboard and /dashboard/appel (for Teleop)
Added /dashboard where a teleoperateur can choose a client to call and a script to use.
Added /dashboard/appel which fills in the chosen script, a list of all products, and all the info regarding a client, with the choice for the teleoperateur to fill in the selected product, it's quantity and the result of the call.
Added /dashboard/sauvegarder-appel to save the info provided in /dashboard/appel.
Added TeleoperateurController.
Added DashboardController@dashboardRouting which now handles all /dashboard routing. this function responds with a view depending on account type, except for Teleoperateur, where it instantiates a object of type TeleoperateurController and calls the function callSetup.
Added model Call and migration create_calls_table which store the data provided by a teleoperateur in /dashboard/appel.