Temporary fix for /dashboard
The manager's /dashboard breaks if the team didn't have any calls for the past 7 days. So for now, it is fixed to the week of 15/03/2022, as this week had several calls. this means that the /dashboard page no longer reflects accurately what happened during the week. it still needs a lot of work and it's very much WIP.
Manager account creation is still (currently) disabled.
Added my manager and teleoperateur login details in the login page as "Demo accounts", Yes they can do destructive things but I still have backup of the DB.
Added the Rapport .docx and the entire DB .sql files into the root folder of the project files.